Use epoll to poll the event module eventfd
eventfds will always signal either POLLIN or POLLOUT, so mask
out the POLLOUT by wrapping it in an epoll fd for those applications
that try to poll on both read and write readiness.

Of course, this means that POLLOUT will never be signalled, so the
application should check both read and write events when POLLIN is signalled.