Don't expose appsink preroll buffers to V4L2
Appsink holds a reference to the preroll buffer until the
pipeline enters a non-playing state. As a result, this buffer
will never be updated.  This can lead to a V4L2 application
waiting forever for this buffer to dequeue.

Requesting an extra buffer from GStreamer to hold the preroll and
exposing the remaining buffers to the application avoids this problem.

The preroll handler is also a good place to do the one time initialization
needed for playback instead of checking a condition on every frame.