Add input buffering to ensure necessary output buffers are available
When decoding in Chromium, if the output buffer format becomes available
too soon, there will not be enough processed input data available to fill
all of the requested output buffers.  This will mean that
the plugin will not be able to provide mmaps/dmabufs for all of the output
buffers as the buffers are still held inside the GStreamer pipeline.
This patch adds some input buffering to avoid this condition though it
may not be enough if the number of buffers is larger than about 10.

This is a temporary workaround until a more robust solution is available.